LPS-76837 Move to taglib module
